Add pool_recycle and pool_size to mysql sqlalchemy
Add pool_recycle and pool_size to mysql sqlalchemy backend.

MySQL has a setting `wait_timeout` which will drop connections after being idle for a certain time. This setting is defaults to 8 hours but some distributions default it to 10 minutes. If a connection is dropped by MySQL without SQLAlchemy expecting it it will cause the client to raise an exception with the message (2006, 'MySQL server has gone away'). Setting `pool_recycle` to less than `wait_timeout` is a common way in SQLAlchemy to work around this as it will cause SQLAlchemy to replace any connections older than this value.