Add shell command to Django backend
This PR adds a `shell` command to `apistar.backends.django_orm.commands`
to make it easier to access a Django-ready shell.

I had a hard time writing a test that asserted that a Python shell got
launched when the command was executed, but I can take another crack at
adding such a test if you'd like.