Update cli.py
It seems as if the theme parameter was not passed in. As a result, the `index.html` does not change along with the JS/CSS resources that are made at the bottom of this particular routine. Not sure if there are other issues in this function, but this fixed theming for me.